Potions now decrease as you use them. doesnt really do anything though.

+++ Assets/Scripts/MenuScripts/InventoryPanel.cs
@@ -9,7 +9,7 @@ public class InventoryPanel : MonoBehaviour {
 
 
 	string ItemName;
-	int quantity;
+	float quantity;
 
 	private GameObject Inventory;
 	private Transform Item1;//Item 1
